Create setup repository service
